+++ src/main/java/ru/nsu/contactproxy/backend/repositories/entities/RoleAttachedField.java
@@ -17,15 +17,15 @@ import javax.annotation.Generated;
 public class RoleAttachedField {
 
   @JsonProperty("id")
-  private Integer id;
+  private Long id;
 
   @JsonProperty("roleId")
-  private Integer roleId;
+  private Long roleId;
 
   @JsonProperty("userFieldTypeId")
-  private Integer userFieldTypeId;
+  private Long userFieldTypeId;
 
-  public RoleAttachedField id(Integer id) {
+  public RoleAttachedField id(Long id) {
     this.id = id;
     return this;
   }
@@ -36,15 +36,15 @@ public class RoleAttachedField {
   */
   @NotNull 
   @Schema(name = "id", example = "123", required = true)
-  public Integer getId() {
+  public Long getId() {
     return id;
   }
 
-  public void setId(Integer id) {
+  public void setId(Long id) {
     this.id = id;
   }
 
-  public RoleAttachedField roleId(Integer roleId) {
+  public RoleAttachedField roleId(Long roleId) {
     this.roleId = roleId;
     return this;
   }
@@ -55,15 +55,15 @@ public class RoleAttachedField {
   */
   @NotNull 
   @Schema(name = "roleId", example = "123", required = true)
-  public Integer getRoleId() {
+  public Long getRoleId() {
     return roleId;
   }
 
-  public void setRoleId(Integer roleId) {
+  public void setRoleId(Long roleId) {
     this.roleId = roleId;
   }
 
-  public RoleAttachedField userFieldTypeId(Integer userFieldTypeId) {
+  public RoleAttachedField userFieldTypeId(Long userFieldTypeId) {
     this.userFieldTypeId = userFieldTypeId;
     return this;
   }
@@ -74,11 +74,11 @@ public class RoleAttachedField {
   */
   @NotNull 
   @Schema(name = "userFieldTypeId", example = "123", required = true)
-  public Integer getUserFieldTypeId() {
+  public Long getUserFieldTypeId() {
     return userFieldTypeId;
   }
 
-  public void setUserFieldTypeId(Integer userFieldTypeId) {
+  public void setUserFieldTypeId(Long userFieldTypeId) {
     this.userFieldTypeId = userFieldTypeId;
   }
